+++ src/renderengine/creator/index.js
@@ -9,38 +9,8 @@ import Obj from './Obj'
 import Polygon from './Polygon'
 import Primitive from './Primitive'
 import Rect from './Rect'
+import RoundRect from './RoundRect'
 import Stripes from './Stripes'
-
-class RoundRect extends Obj {
-  getName = 'Rounded Rectangle'
-
-  list = [
-    // { mY:1 },
-    // { mX:1, height: {a:1} },
-    // { mX:1, height: {a:1}, fromBottom:true },
-    {
-      minX: 3,
-      minY: 4,
-      list: [
-        { name: 'Dot', clear: true },
-        { name: 'Dot', fX: true, clear: true },
-        { name: 'Dot', fY: true, clear: true },
-        { name: 'Dot', fX: true, fY: true, clear: true },
-      ],
-    },
-    {
-      minX: 4,
-      minY: 3,
-      list: [
-        { name: 'Dot', clear: true },
-        { name: 'Dot', fX: true, clear: true },
-        { name: 'Dot', fY: true, clear: true },
-        { name: 'Dot', fX: true, fY: true, clear: true },
-      ],
-    },
-    {},
-  ]
-}
 
 class Grid extends Obj {
   getName = 'Grid'
